add readme and current solution for longest consecutive sequence
The longest consecutive sequence solution added here was not my first draft, just what I had when I created this repo. The main problem with the current solution is that it takes too long. This is likely because I am walking the `sorted_nums` list for each value in the `nums` list - breaching linear time complexity.