completely refactor solution for longest consecutive sequece to avoide nested loops
New solution to longest consecutive sequence basically creates a list of poteltial boolean flags and flags a position in the list which matches a number in the provided `nums` list. Sequential `true` values in the boolean list will represent consequtive sequences in the original list.

The problem here is I am creating a stack overflow with the `sorted_nums` boolean list.